CheckOut not done (need update stockQuantity and update point 4 user)

+++ src/main/java/com/mycompany/isp392/cart/CartDTO.java
@@ -45,7 +45,7 @@ public class CartDTO {
         this.cartID = cartID;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(double totalPrice) {
         this.totalPrice = totalPrice;
     }
